+++ packages/durable-event-iterator/src/client/index.ts
@@ -1,0 +1,1 @@
+export * from './event-iterator'
