+++ packages/durable-event-iterator/src/index.ts
@@ -1,4 +1,3 @@
 export * from './builder'
-export * from './client'
-export * from './durable-event-iterator/object'
-export * from './server'
+export * from './event-iterator'
+export * from './object'
